Maj tortue attention tortue pas ouf à refaire

+++ src/alphabet/Avancer.java
@@ -3,18 +3,18 @@ import java.util.LinkedList;
 
 public class Avancer extends Mouvement {
 	private static final Character REPRESENTATION_BASE='f';
-	private static final double VALEUR_BASE=10;
+	private static final int VALEUR_BASE=10;
 	
-	public Avancer(Character representation, LinkedList<MembreAlpha> evolution ,double valeur )  {
+	public Avancer(Character representation, LinkedList<MembreAlpha> evolution ,int valeur )  {
 		super(representation,evolution,valeur);
 	}
-	public Avancer(Character representation,double valeur) {
+	public Avancer(Character representation,int valeur) {
 		super(representation,valeur);
 	}
 	public Avancer() {
 		this(REPRESENTATION_BASE,VALEUR_BASE);
 	}
-	public Avancer(double valeur) {
+	public Avancer(int valeur) {
 		this(REPRESENTATION_BASE,valeur);
 	}
