+++ src/alphabet/DessinerAvancer.java
@@ -4,14 +4,14 @@ import java.util.LinkedList;
 
 public class DessinerAvancer extends Mouvement {
 	private static final Character REPRESENTATION_BASE='F';
-	private static final double VALEUR_BASE=10;
+	private static final int VALEUR_BASE=10;
 	
-	public DessinerAvancer(Character representation, LinkedList<MembreAlpha> evolution, double valeur) {
+	public DessinerAvancer(Character representation, LinkedList<MembreAlpha> evolution, int valeur) {
 		super(representation, evolution, valeur);
 		this.dessine=true;
 	}
 
-	public DessinerAvancer(Character representation, double valeur) {
+	public DessinerAvancer(Character representation, int valeur) {
 		super(representation, valeur);
 		this.dessine=true;
 	}
@@ -19,7 +19,7 @@ public class DessinerAvancer extends Mouvement {
 	public DessinerAvancer() {
 		this(REPRESENTATION_BASE,VALEUR_BASE);
 	}
-	public DessinerAvancer(double valeur) {
+	public DessinerAvancer(int valeur) {
 		this(REPRESENTATION_BASE,valeur);
 	}
